Bicleaner support + fixes (#13)
SacreBLEU is a regular importer now and evaluation is not limited to sacrebleu datasets.
fixes

Added bicleaner-ai and bicleaner filtering (one or another based on available pretrained language packs).
fixes


Added script to find all datasets based on language pair and importer type, ready to use in config
fixes


Fixed conda environment activation to be reproducible on GCP

Other minor reproducibility fixes

+++ pipeline/data/importers/corpus/mtdata.sh
@@ -16,18 +16,25 @@ trg=$2
 dir=$3
 dataset=$4
 
+test -v WORKDIR
+
+source "${WORKDIR}/pipeline/setup/activate-python.sh"
+
 src_iso=$(python -c "from mtdata.iso import iso3_code; print(iso3_code('${src}', fail_error=True))")
 trg_iso=$(python -c "from mtdata.iso import iso3_code; print(iso3_code('${trg}', fail_error=True))")
 
-mtdata get -l "${src}-${trg}" -tr "${dataset}" -o "${dir}"
+if [ ! -e "${dir}/${dataset}.${TRG}" ]; then
+  mtdata get -l "${src}-${trg}" -tr "${dataset}" -o "${dir}"
 
-for f in "${dir}"/train-parts/*."${src_iso}"; do
-  mv "${f}" "${dir}/${dataset}.${SRC}"
-done
-for f in "${dir}"/train-parts/*."${trg_iso}"; do
-  mv "${f}" "${dir}/${dataset}.${TRG}"
-done
+  for f in "${dir}"/train-parts/*."${src_iso}"; do
+    mv "${f}" "${dir}/${dataset}.${SRC}"
+  done
+  for f in "${dir}"/train-parts/*."${trg_iso}"; do
+    mv "${f}" "${dir}/${dataset}.${TRG}"
+  done
 
-rm -rf "${dir}/train-parts"
+  rm -rf "${dir}/train-parts"
+fi
+
 
 echo "###### Done: Downloading mtdata corpus"
